+++ boop/__init__.py
@@ -3,7 +3,7 @@
 #
 """PyBoop - A fun 2d Game Engine built on top of Pyglet."""
 
-version='0.0a11'
+version='0.0a12'
 
 from . import component
 from .boopwindow import BoopWindow