+++ boop/drawables.py
@@ -13,9 +13,10 @@ from .drawable import Drawable, DEBUG_DRAWABLES
 class DrawWrapper(Drawable):
     """Wraps a draw_* routine (or any other callable) to the Drawable
        protocol."""
+    ### FIXME we should provide a way to implement dimensions and other things here.
 
     def __init__(self, display, drawtool, *args, **kwargs):
-        Drawable.__init__(self, display)
+        Drawable.__init__(sef, display)
         self.drawtool = drawtool
         self.dt_args = args
         self.dt_kwargs = kwargs
@@ -26,6 +27,7 @@ class DrawWrapper(Drawable):
 
 
 class DraggableDrawableMixin(DragMixin):
+    """Mix this with a Drawable to allow the drawable to be dragged and dropped."""
     _ddm_offset = (0, 0)
 
     def can_drag(self, state, x, y):
@@ -46,43 +48,31 @@ class DraggableDrawableMixin(DragMixin):
 
 # Proof of concept, probably not a good thing to actually use.
 class Clear(Drawable):
+    """This is a proof of concept, and should probably not be used (put a window.clear in the render_* for the Scene or decorative Drawable)."""
     def do_render(self, window):
         window.clear()
 
 
 class Image(Drawable):
-    def __init__(self, display, imgobj, position=(0, 0)):
+    """Wraps pyglet's Sprite in Drawable protocol. Not effecient if there are many sprites."""
+    ### FIXME implement BAG OF IMAGE type which has dirtyness setting and only reloads sprites into a batch if they are dirty.
+    def __init__(self, display, imgobj, position=(0, 0), z=0.0):
         Drawable.__init__(self, display)
         self.spr = pyglet.sprite.Sprite(imgobj)
-        self.spr.z = 0.0
+        self.spr.z = float(z)
         self.setpos(*position)
-
-    def getpos(self):
-        pos = Drawable.getpos(self)
-        return pos
-
-    def setpos(self, x, y):
-        Drawable.setpos(self, x, y)
-        # self.spr.position = (x, y)
 
     def rotate(self, rotation):
         self.spr.rotation = float(rotation)
         self.setpos(*self.spr.position)
-        # fixme calculate new rectangle based on rotated points (width and
-        # height aren't fixed by pyglet) note that it rotates around 0,0
-        # regardless of anchor, which may make the transformation
-        # much easier to deal with.
 
-        # self.width = self.spr.width
-        # self.height = self.spr.height
+    def setscale(self, scale):
+        self.spr.scale = scale
 
     def getsize(self):
         return self.spr.width, self.spr.height
 
     def do_render(self, display):
-        # self.spr.x = float(self.position[0])
-        # self.spr.y = float(self.position[1])
-        # self.spr.z = 0.0
         GL.glEnable(GL.GL_BLEND)
         GL.glBlendFunc(GL.GL_SRC_ALPHA, GL.GL_ONE_MINUS_SRC_ALPHA)
         self.spr.draw()
@@ -98,9 +88,8 @@ class Image(Drawable):
             # as drawable position)
             drawtools.draw_crosshair(pos[0], pos[1], color=(0.0, 1.0, 1.0))
 
-
-
 class Backdrop(Image):
+    """A deprecated proof of concept that draws a large image the size of the window it's placed in."""
     def __init__(self, display, imgobj):
         Image.__init__(self, display, imgobj, position=(0, 0))
         scale = 0
@@ -121,6 +110,7 @@ class Backdrop(Image):
 
 
 class Label(Drawable):
+    """Wraps pyglet's text object class in Drawable protocol."""
     def __init__(self,
                  display,
                  font,
@@ -158,6 +148,7 @@ class Label(Drawable):
 
 
 class FadeMixin(object):
+    """Mix this with your Drawable which supports alpha in order to fade it in and out using tick events."""
     FADE_ST_ON, FADE_ST_OFF, FADE_ST_FO, FADE_ST_FI = range(4)
 
     def __init__(self, *args, **kwargs):
@@ -187,12 +178,14 @@ class FadeMixin(object):
 
 
 class DraggableImage(Image, DraggableDrawableMixin):
+    """An Image that's also Draggable. Deprecated proof of concept that may not be useful in practice."""
     def __init__(self, window, image):
         DraggableDrawableMixin.__init__(self)
         Image.__init__(self, window, image)
 
 
 class ClockDisplay(Drawable):
+    """Simple display for FPS counter."""
     #fixme apparently this is inaccurate according to docs. newer pyglet has fix.
     # fixme map normal drawable parms to this (fps.label is a regular label object).
     def __init__(self, window, *args, **kwargs):